requirements.txt
streamlit>=1.24.0
openai>=0.27.0
python-dotenv>=1.0.0
python-docx>=0.8.11
pdfkit>=1.0.0
google-api-python-client>=2.80.0
google-auth>=2.23.4
ffmpeg-python>=0.2.0 